[Fixed 생산량 예측모들 함수 인자 float 변경]

+++ backend/utils/singletone.py
@@ -34,7 +34,7 @@ class ModelSingleTone(SingletonInstane):
     def get_lgb_model(self):
         return self.lgb_model
     
-    def get_crop_production(self, crop_id:int, area:int):
+    def get_crop_production(self, crop_id:float, area:float):
         y_column = ["output"]
         crop_id = [crop_id]
         area = [area]
